add debug flag
