+++ src/utils/utils.h
@@ -1,28 +1,88 @@
 #ifndef MenuUtils_H
 #define MenuUtils_H
-
 #include <Arduino.h>
 
-void substring(const char* str, uint8_t start, uint8_t size, char* substr);
+#include "constants.h"
 
-void concat(const char* first, char second, const char* third, char* result);
+void substring(const char* str, uint8_t start, uint8_t size, char* substr) {
+    strncpy(substr, str + start, size);
+    substr[size] = '\0';
+}
 
-void concat(const char* first, char second, char* result);
+void concat(const char* first, char second, const char* third, char* result) {
+    size_t len1 = strlen(first);
+    size_t len3 = strlen(third);
 
-void concat(const char* first, const char* second, char* result);
+    memcpy(result, first, len1);
+    result[len1] = second;
+    memcpy(result + len1 + 1, third, len3);
+    result[len1 + 1 + len3] = '\0';
+}
 
-void remove(char* str, uint8_t index, uint8_t count);
+void concat(const char* first, char second, char* result) {
+    size_t len1 = strlen(first);
+    memcpy(result, first, len1);
+    result[len1] = second;
+    result[len1 + 1] = '\0';
+}
 
-long mapProgress(uint16_t progress, long minValue, long maxValue);
+void concat(const char* first, const char* second, char* result) {
+    strcpy(result, first);
+    strcat(result, second);
+}
 
-float mapProgress(uint16_t progress, float minValue, float maxValue);
+void remove(char* str, uint8_t index, uint8_t count) {
+    uint8_t len = strlen(str);
+    if (index + count > len) {
+        count = len - index;
+    }
+    memmove(str + index, str + index + count, len - count - index + 1);
+}
 
-void printCmd(const __FlashStringHelper* command);
+long mapProgress(uint16_t progress, long minValue, long maxValue) {
+    // Map the progress value to a new range (minValue to maxValue)
+    return map(progress, MIN_PROGRESS, MAX_PROGRESS, minValue, maxValue);
+}
 
-void printCmd(const __FlashStringHelper* command, const char value);
+float mapProgress(uint16_t progress, float minValue, float maxValue) {
+    // Normalize the progress value and map it to the specified floating-point
+    // range
+    return static_cast<float>(progress) / MAX_PROGRESS * (maxValue - minValue) +
+           minValue;
+}
 
-void printCmd(const __FlashStringHelper* command, const uint8_t value);
+void printCmd(const __FlashStringHelper* command) {
+#ifdef DEBUG
+    Serial.print(F("#CMD# "));
+    Serial.println(command);
+#endif
+}
 
-void printCmd(const __FlashStringHelper* command, const char* value);
+void printCmd(const __FlashStringHelper* command, const char value) {
+#ifdef DEBUG
+    Serial.print(F("#CMD# "));
+    Serial.print(command);
+    Serial.print(F("="));
+    Serial.println(value);
+#endif
+}
+
+void printCmd(const __FlashStringHelper* command, const uint8_t value) {
+#ifdef DEBUG
+    Serial.print(F("#CMD# "));
+    Serial.print(command);
+    Serial.print(F("="));
+    Serial.println(value, DEC);
+#endif
+}
+
+void printCmd(const __FlashStringHelper* command, const char* value) {
+#ifdef DEBUG
+    Serial.print(F("#CMD# "));
+    Serial.print(command);
+    Serial.print(F("="));
+    Serial.println(value);
+#endif
+}
 
 #endif
